Remove FlatSpatialPooler
